llama3-rag-observability/requirements.txt
qdrant-client==1.9.1
llama-index==0.10.35
llama-index-llms-ollama==0.1.3
llama-index-vector-stores-qdrant==0.2.8
llama-index-readers-file==0.1.21
llama-index-embeddings-ollama==0.1.2
llama-index-callbacks-arize-phoenix>0.1.2